chore: logs and constants
